usb: pd: Register typec partner in case AAA is connected
Currently PD driver is not registering typec partner if
analog audio adapter(AAA) is connected to the device. Because of
this, userspace will not be notified of the connected adapter
and there will be no notification seen on UI. This change
registers typec partner in case AAA is connected and unregister
it upon disconnect.

Change-Id: Iffd9a43ded068047dd175ccf66de43160da4daaf
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>