arm64: setup: Sony: Fix false positives of offline charging mode
The offline charging mode parameter should be set only when the
wall charger or USB-PC cable is connected to a turned off device or
the device is turned off with the wall charger or USB-PC cable
connected, for this the bootloader sets several values ​​for the startup
parameter:
  - 0x00004000 for wall charger,
  - 0x00000004 for USB-PC.

In the case of using a bitwise AND with a value of 0x4004 to check the
startup parameter, we will receive false positives and will jump into
offline charging mode when this should not happen, for example, when
rebooting with a connected USB-PC cable, the bootloader will set the
startup parameter to 0x00008004, which with a bitwise logical AND will
return the result 0x4 (true). The same behavior will be observed during
rebooting with the wall charger connected, the bootloader will set the
startup parameter to 0x00004000, which with a bitwise logical AND will
return the result 0x4000 (true), etc.

To prevent this and restore the normal functioning of the device reboot
during various restart reasons with a connected wall charger or USB
cable, correct the check of the startup flag so that the offline
charging parameter is set only in two cases described at the beginning.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>