soc: qcom: llcc-sm8150: Update the system cache table
Regenerate the System Cache Table to match the current kernel version.

Additional changes contained in this commit:
1. The sm8150-llcc driver init and exit functions are calling
platform_driver_register and platform_driver_unregister and
doing nothing else. This same as that of the module_platform_driver,
remove this init and exit functions and use the module_platform_driver
instead.

2. Add SPDX-License-Identifier mark to source file.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>