misc: sony_ext_uim_ctrl: Introduce uim2_detect_legacy property
Some platform only uses one of the GPIOs and sets its value to HIGH for
second SIM slot detection. For this we used several ifdef macros, which
cluttered up the code quite a lot due to the large number of platforms
using the old second SIM slot detection method, in order to make it a
little cleaner add the bool uim2_detect_legacy property and read the
value of it from the devicetree if it is required by some platforms.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>